+++ .github/workflows/tests.yml
@@ -1,0 +1,30 @@
+name: Tests
+
+on: [push, pull_request, workflow_dispatch]
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 2
+      matrix:
+        python-version: [3.9]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Set up tox env
+      run: |
+        pver=${{ matrix.python-version }}
+        tox_env="-epy${pver/./}"
+        echo tox_env
+        echo TOX_ENV=$tox_env >> $GITHUB_ENV
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Test using tox envs
+      run: |
+        tox ${{ env.TOX_ENV }}
