+++ quantum_glasses/__init__.py
@@ -1,0 +1,3 @@
+"""Main module."""
+
+from .quantum_glasses import QuantumGlasses
