+++ requirements.txt
@@ -1,25 +1,4 @@
-contourpy==1.1.0
-cycler==0.11.0
-dill==0.3.7
-exceptiongroup==1.1.2
-fonttools==4.42.0
-iniconfig==2.0.0
-kiwisolver==1.4.4
-matplotlib==3.7.2
-mpmath==1.3.0
-numpy==1.25.2
-packaging==23.1
-Pillow==10.0.0
-pluggy==1.2.0
-ply==3.11
-psutil==5.9.5
-pyparsing==3.0.9
+fire==0.5.0
+qiskit==0.44.1
+qiskit[visualization]==0.44.1
 pytest==7.4.0
-python-dateutil==2.8.2
-qiskit==0.44.0
-qiskit-terra==0.25.0
-rustworkx==0.13.1
-six==1.16.0
-stevedore==5.1.0
-sympy==1.12
-typing_extensions==4.7.1
