+++ tox.ini
@@ -1,0 +1,16 @@
+[tox]
+minversion = 3.9
+envlist = py39
+skipsdist = True
+
+[testenv]
+usedevelop = true
+install_command = pip install -U {opts} {packages}
+setenv =
+  VIRTUAL_ENV={envdir}
+  LANGUAGE=en_US
+  LC_ALL=en_US.utf-8
+deps = -r requirements.txt
+commands =
+  pip check
+  python -m unittest -v
